requirements.txt
aioredis==1.2.0
asgiref==2.3.2
async-timeout==2.0.1
attrs==18.2.0
autobahn==18.11.1
Automat==0.7.0
channels==2.1.5
channels-redis==2.3.1
Click==7.0
constantly==15.1.0
cycler==0.10.0
daphne==2.2.3
Django==1.11
graphviz==0.10.1
gunicorn==19.9.0
h11==0.8.1
hiredis==0.2.0
httptools==0.0.11
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
kiwisolver==1.0.1
matplotlib==3.0.1
memory-profiler==0.54.0
msgpack==0.5.6
numpy==1.15.4
objgraph==3.4.0
psutil==5.4.8
PyHamcrest==1.9.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
Twisted==18.9.0
txaio==18.8.1
uvicorn==0.3.20
uvloop==0.11.3
websocket-client==0.54.0
websockets==7.0
zope.interface==4.6.0
